backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.15",
    "ruff>=0.14.8",
]

[tool.pyright]
typeCheckingMode = "off"